requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-redis==5.4.0
django-environ==0.9.0
django-ckeditor==6.7.3

pillow==11.2.1

uvicorn==0.34.2

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.9.0

psycopg2==2.9.4

celery==5.5.3
django-celery-results==2.6.0
django-celery-beat==2.8.1
